+++ src/constants/network.ts
@@ -1,19 +1,5 @@
-import { Aptos, AptosConfig, Network } from '@aptos-labs/ts-sdk'
+import { Network } from '@aptos-labs/ts-sdk'
 import { AptosPriceServiceConnection } from '@pythnetwork/pyth-aptos-js'
-import { AuthConfig, authExchange, AuthUtilities } from '@urql/exchange-auth'
-import { cacheExchange, Client, CombinedError, createClient, errorExchange, fetchExchange, Operation } from 'urql'
-
-// store copies of the client so they don't need to be rebuilt when fetched
-const defaultMainnetClient = new Aptos(new AptosConfig({ network: Network.MAINNET }))
-const defaultTestnetClient = new Aptos(new AptosConfig({ network: Network.TESTNET }))
-// https://aptos.testnet.suzuka.movementlabs.xyz/v1
-const defaultMovementTestnetClient = new Aptos(
-  new AptosConfig({
-    network: Network.CUSTOM,
-    fullnode: 'https://aptos.testnet.suzuka.movementlabs.xyz/v1',
-    indexer: 'https://indexer.testnet.suzuka.movementlabs.xyz/v1/graphql',
-  }),
-)
 
 const mainnetPythClient = new AptosPriceServiceConnection(`https://hermes.pyth.network`, {
   priceFeedRequestConfig: {
@@ -31,28 +17,6 @@ const testnetPythClient = new AptosPriceServiceConnection(`https://hermes-beta.p
     binary: true,
   },
 })
-
-/**
- * Get an Aptos client for a network
- * @param network the network to use, if not specific = "mainnet"
- * @returns a useable aptos client
- */
-export const defaultAptosClient = (network: Network | string | string = Network.MAINNET, nodeURI?: string): Aptos => {
-  if (nodeURI !== undefined) {
-    return new Aptos(new AptosConfig({ network: getNetwork(network), fullnode: nodeURI }))
-  }
-  switch (network) {
-    case Network.MAINNET:
-      return defaultMainnetClient
-    // TODO make it work for movement-mainnet (network custom needs a seperate switch somehow)
-    case Network.CUSTOM:
-      return defaultMovementTestnetClient
-    case 'movement-testnet':
-      return defaultMovementTestnetClient
-    default:
-      return defaultTestnetClient
-  }
-}
 
 /**
  * Get a Pyth price service connection
@@ -77,75 +41,4 @@ export const getNetwork = (network: Network | string): Network => {
     }
   }
   return network
-}
-
-export const graphqlClientWithUri = (gqlURI: string, API_KEY?: string): Client => {
-  const exchanges = [
-    fetchExchange,
-    cacheExchange,
-    errorExchange({
-      onError: (error) => {
-        console.error('GraphQL Error:', error)
-      },
-    }),
-  ]
-
-  if (API_KEY) {
-    exchanges.push(
-      authExchange(async (utils: AuthUtilities): Promise<AuthConfig> => {
-        return {
-          addAuthToOperation(operation: Operation): Operation {
-            return utils.appendHeaders(operation, {
-              Authorization: `Bearer ${API_KEY}`,
-            })
-          },
-          didAuthError: (error: CombinedError): boolean => {
-            return error.response.status === 401
-          },
-          refreshAuth: async (): Promise<void> => {},
-        }
-      }),
-    )
-  }
-
-  return createClient({
-    url: gqlURI,
-    exchanges,
-  })
-}
-
-// TODO: make this not fixed to aptos testnet
-export const defaultAptosGraphqlClient = (network: string, API_KEY?: string): Client => {
-  // TODO: UPDATE WITH NEW API ENDPOINTS
-  let uriStr = 'https://api.testnet.aptoslabs.com/v1/graphql'
-  if (network == 'testnet') {
-    uriStr = 'https://api.testnet.aptoslabs.com/v1/graphql'
-  } else if (network == 'mainnet') {
-    uriStr = 'https://api.mainnet.aptoslabs.com/v1/graphql'
-  } else if (network == 'movement-testnet') {
-    uriStr = 'https://indexer.testnet.porto.movementnetwork.xyz/v1/graphql'
-  }
-  return graphqlClientWithUri(uriStr, API_KEY)
-}
-
-export const defaultMirageGraphqlClient = (network: string): Client => {
-  // TODO: UPDATE WITH NEW API ENDPOINTS
-  let uriStr = ''
-  if (network == 'testnet') {
-    uriStr = 'https://api-movement-testnet.mirage.money/v1/graphql'
-  } else {
-    uriStr = 'https://api-movement-testnet.mirage.money/v1/graphql'
-  }
-  return createClient({
-    url: uriStr,
-    exchanges: [
-      fetchExchange,
-      cacheExchange,
-      errorExchange({
-        onError: (error) => {
-          console.error('GraphQL Error:', error)
-        },
-      }),
-    ],
-  })
 }
