+++ src/entities/index.ts
@@ -1,5 +1,7 @@
 export * from './asset'
 export * from './assetEntities'
+export * from './baseEntities'
 export * from './market'
+export * from './mirage_asset'
 export * from './rebase'
 export * from './vault'
