+++ src/entities/market/index.ts
@@ -1,3 +1,4 @@
 export * from './limitOrder'
 export * from './market'
+export * from './marketEntities'
 export * from './position'
