+++ src/views/index.ts
@@ -1,4 +1,6 @@
 export * from './accountViews'
+export * from './baseViews'
 export * from './marketViews'
+export * from './referralViews'
 export * from './testnetViews'
 export * from './vaultViews'
